Refactor whole folder flow_counter
