+++ tests/bgp/test_bgp_speaker.py
@@ -15,8 +15,7 @@ from tests.ptf_runner import ptf_runner
 from tests.common.utilities import wait_tcp_connection
 from tests.common.helpers.assertions import pytest_require
 from tests.common.utilities import wait_until
-from tests.flow_counter.flow_counter_utils import RouteFlowCounterTestContext, \
-                                                  is_route_flow_counter_supported   # noqa F401
+from tests.common.flow_counter import RouteFlowCounterTestContext  # noqa F401
 
 
 pytestmark = [
